# specs/001-outbound-callcenter-poc/contracts/api.yaml
openapi: 3.0.3
info:
  title: Contact Center POC API
  description: >
    ASP.NET Core Web API for the Outbound Call Center POC.
    Orchestrates outbound phone calls via Azure Communication Services
    and bridges audio to Azure OpenAI Realtime API for AI-driven conversations.
    Supports campaign management, real-time sentiment analysis, multi-number calling,
    and call history with analytics.
  version: 3.0.0
servers:
  - url: https://{host}
    description: API base URL (ngrok tunnel or Azure deployment)
    variables:
      host:
        default: localhost:5001
paths:
...
  /api/CallHistory/{callConnectionId}:
    get:
      operationId: getCallDetail
      summary: Get detailed call record
      description: >
        Returns the full call record including transcript with per-segment
        sentiment, speaker timeline, and aggregate analytics (sentiment
        breakdown, talk-time ratio).
      tags: [CallHistory]
      parameters:
        - name: callConnectionId
          in: path
          required: true
          schema:
            type: string
          description: The ACS call connection ID of the completed call
      responses:
        '200':
          description: Call detail with full transcript and analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallRecord'
        '404':
          description: No call record found with this ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/CallHistory/{callConnectionId}/recording:
    get:
      operationId: getCallRecording
      summary: Download call recording audio
      description: >
        Streams the MP3 recording for a completed call from Azure Blob Storage.
        Returns the audio file with Content-Type audio/mpeg and supports HTTP
        range requests for seeking in the browser audio player.
      tags: [CallHistory]
      parameters:
        - name: callConnectionId
          in: path
          required: true
          schema:
            type: string
          description: The ACS call connection ID
      responses:
        '200':
          description: Recording audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '404':
          description: No recording found for this call
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/CallHistory/{callConnectionId}/transcribe:
    post:
      operationId: transcribeCallRecording
      summary: Transcribe a call recording
      description: >
        Downloads the recording audio from Blob Storage, sends it to Azure
        OpenAI for transcription, and persists the transcript text into the
        call history record. Returns the updated CallRecord with the transcript.
      tags: [CallHistory]
      parameters:
        - name: callConnectionId
          in: path
          required: true
          schema:
            type: string
          description: The ACS call connection ID
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, re-transcribe even if a transcript already exists
      responses:
        '200':
          description: Transcription completed, updated call record returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallRecord'
        '404':
          description: No call record or recording found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error category
        message:
          type: string
          description: Human-readable error detail
    TranscriptEntry:
      type: object
      description: Streamed via SignalR (TranscriptHub) and persisted in CallRecord
      properties:
        callConnectionId:
          type: string
        speaker:
          type: string
          enum: [AI, Recipient]
        text:
          type: string
        sentiment:
          $ref: '#/components/schemas/SentimentResult'
        emotion:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/EmotionResult'
          description: Per-segment emotion (Phase 22). Null for pre-Phase-22 entries.
        timestamp:
          type: string
          format: date-time
...
    SentimentResult:
      type: object
      properties:
        label:
          type: string
          enum: [Positive, Neutral, Negative]
          description: Sentiment classification
        confidence:
          type: number
          format: float
          description: Confidence score (0.0–1.0)
...
    CallRecord:
      type: object
      description: Full call record with transcript and analytics
      properties:
        callConnectionId:
          type: string
        phoneNumber:
          type: string
        contactName:
          type: string
          nullable: true
          description: Optional contact name (Phase 15)
        campaignId:
          type: string
          nullable: true
        campaignTitle:
          type: string
          nullable: true
        prompt:
          type: string
        duration:
          type: string
          description: Call duration (HH:mm:ss format)
        overallSentiment:
          type: string
          enum: [Positive, Neutral, Negative]
        recordingId:
          type: string
          nullable: true
          description: ACS recording ID (Phase 15)
        hasRecording:
          type: boolean
          description: True if a recording blob exists (Phase 16)
        recordingTranscript:
          type: string
          nullable: true
          description: Transcribed text from the recording (Phase 18)
        recordingTranscribedAt:
          type: string
          format: date-time
          nullable: true
          description: When the recording was transcribed (Phase 18)
        sentimentBreakdown:
          type: object
          properties:
            positivePercent:
              type: number
              format: float
            neutralPercent:
              type: number
              format: float
            negativePercent:
              type: number
              format: float
        talkTimeRatio:
          type: object
          properties:
            aiPercent:
              type: number
              format: float
            recipientPercent:
              type: number
              format: float
        operatorStyleTraits:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/OperatorStyleTraits'
          description: Operator style traits (Phase 22). Null for pre-Phase-22 call records.
        transcriptEntries:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptEntry'
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    EmotionResult:
      type: object
      description: Per-segment emotion classification (Phase 22)
      properties:
        label:
          $ref: '#/components/schemas/EmotionLabel'
        confidence:
          type: number
          format: float
          description: Confidence score (0.0–1.0)
    EmotionLabel:
      type: string
      enum: [Neutral, Happy, Frustrated, Angry, Sad, Anxious]
      description: Emotion classification labels (Phase 22)
    OperatorStyleTraits:
      type: object
      description: Operator style traits computed at end-of-call from operator-side transcript (Phase 22)
      properties:
        empathy:
          type: number
          format: float
          description: Empathy score (0.0–1.0)
        energy:
          type: number
          format: float
          description: Energy score (0.0–1.0)
        computedAt:
          type: string
          format: date-time
          description: UTC timestamp when the traits were computed
...